requirements.txt
torch==2.9.1
torchvision==0.24.1
wandb==0.23.1
pyyaml==6.0.2